requirements/base.txt
Django>=1.10
Unipath>=1.0
psycopg2
django-extensions
flake8
django_compressor
dj-database-url
django-decouple
djangorestframework
django-filter
pygraphviz
django-jinja
django-mptt
raven
django-localflavor
djangorestframework-jwt
python-dateutil
pymongo
requests